A background cell carries its page, so a tile from the other page is the tile that builds
A layer-2 cell was one byte, the page fixed per background by its stream's
address (§10a). The drawer offered all 0x200 BG tiles anyway, so a page-1
tile stamped onto a page-0 background showed as 0x1A5 on the Background
canvas and composed and built as 0x0A5 — the probe that found it: canvas
shows=0x1A5, stored byte=0xA5, scene tile=0x0A5. Half the BG definitions were
unreachable, and the half that was not lied about it.

Rom.BgTilemaps holds whole tile numbers now, page in bit 8. The project keeps
BgTilemap as it was and gains a BgTilemapPages plane beside it; a file saved
before this has none and hydrates against the stream's own plane — per tile
from a custom stream, the address page across a vanilla one — which is what
those cells meant. ToCustomPlanes takes a page per tile (the single-page call
is a wrapper, and LM's five blank extra rows take the plane's own page so a
vanilla background's colours stay uniform), and WriteBackground passes the
real plane through, so the tile that was painted is the tile that builds.

A base without the custom-background hook has one page, from the address, and
can write nothing else. There the editor does what LM does with an
out-of-bank paste (§10c): BgPaintable remaps a drawer tile from the other
page to this page's tile of the same number at PAINT time, the drawer's ring
sits on the tile that will land, the gutter says "page 00 only", and the
build counts any strays it still finds. Named layer-2 tilemaps store two
bytes a cell; a one-byte preset from before takes the level's own pages.

Pinned end to end on $105 (page 0) in a prepped project: stamp 0x1A5, and
the canvas, the level scene, the store, the reopened project and the built
stream's plane all say 0x1A5 while the neighbour keeps page 0. And on a bare
vanilla ROM, 0x1A5 remaps to 0x0A5. One existing test asserted the old
desync and was corrected; the uniform-plane test caught a slip in the
relayout, which was fixed rather than loosened.

+++ src/rom/Rom.cs
@@ -284,11 +284,12 @@ public sealed class Rom
     /// Hydrated from / stashed to ProjectFile.LevelState.Header by LevelSession.</summary>
     public readonly Dictionary<int, byte[]> LevelHeaderOverrides = new();
 
-    /// <summary>Edited layer-2 backgrounds: level → the 0x400 BG Map16 def indices the RLE
-    /// stream decodes to (page NOT applied — it comes from the stream's address, §10a).
-    /// LevelParser.DecodeBgImage prefers these, so an edit reaches the level canvas, the
-    /// Background tab and the built ROM through the one path the base ROM already uses.</summary>
-    public readonly Dictionary<int, byte[]> BgTilemaps = new();
+    /// <summary>Edited layer-2 backgrounds: level → 0x400 whole BG tile numbers (page in bit 8,
+    /// 0x000-0x1FF) at vanilla stride. Whole numbers rather than the stream's low bytes because
+    /// the page is per TILE once a background is custom (§10b); the address-derived page is only
+    /// how a vanilla stream is read in. LevelParser.DecodeBgImage prefers these, so an edit
+    /// reaches the level canvas, the Background tab and the built ROM through one path.</summary>
+    public readonly Dictionary<int, ushort[]> BgTilemaps = new();
 
     /// <summary>Imported layer-3 tilemaps: level → a flat 16-bit map, LM's LT3 file shape
     /// (0x800/0x1000/0x2000 bytes). Replaces vanilla's (level mode, option) pick for that level
